Update to the current version in the Bergen CVS. Most important changes concern the
Standard Tracker:
Added possibility to define where to set the track parameters:
A new boolean static data member AliLevel3::fSetTracks2FirstPoint has been added.
If this is true (default) the tracks are propagated to the first point
on the track (impact point in the TPC), and track parameters are updated
accordingly. In order to define the parameters at the vertex position, you
you can place the call AliLevel3::SetTracks2Vertex, and no propagation will
be done.

Fitting:
Removed obsolute code
Bugfix in the fitting routing in AliL3FitUtilities.
Small change in how the fit range in the padrow plane is estimated.

Compression:
Changed the track parameters in the compression scheme. There are now 5
floats (instead of 6) used to encode the properties of the helix. These
are defined (as before) in AliL3Models.h.