Added the digit reader classes that allow reading of raw packed and unpacked data via the same interface.
Copied the AliHLTTPCClustFinderNew class into the AliHLTTPCClusterFinder
class and modified this to use the new digit reader class(es) for reading
of its input data. The cluster finder can therefore now work both on
the raw packed data (as received from the TPC) and the pre-unpacked data
as delivered by the RawDataUnpackerComponent.

Added the TPC online display ROOT code in the OnlineDisplay subdirectory.