Implementation of local caching in the CDB access framework exploiting the TFile caching functionality.
Local caching is activated in the AliCDBStorage initialisation. It is sufficent
to specify a local cache folder in the AliEn storage URI, e.g.:

AliCDBManager::Instance()->SetDefaultStorage("alien://dbFolder=/alice/simulation/OCDB?cacheFolder=/tmp/LocalCache");

If cache folder is specified AliCDBGrid opens the remote file with the option
"CACHEREAD".

In aliSimulation and AliReconstruction the activation of the "RemoteStorage" is
now obsolete and has been commented out.