Improved ALTRO mapping:  Now does the full conversion from
    <DDL>/<FEC>/<ALTRO>/<Channel>/<Timebin> 
  ->
    FMD<detector><ring>[<sector>,<strip>]-<sample>

and the other way back. Changed all relevant code to use this conversion.
This ensures that the conversion is consistent over all the code. To do
this, I had to implement the AliFMDDigit::SetCount member function.
Not nice, but so be it.  Alternatively, AliFMDRawReader could be a friend
of AliFMDDigit - but that's not nice either. 

Fixed up AliFMDPattern to display the modules correctly.