Major change in the setup of HLT simulation: making CA tracking and merging the default TPC online reconstruction.
In order to also propagate the MC information, the reconstruction
needs to run from digits, which is at the moment done when no
raw data is simulated (AliSimulation::SetWriteRawData not used).
Note: the online reconstruction is chosen at the level of
AliSimulation, not AliReconstruction. It is planned to simulate 
both the HLT chain running on digits and simulated raw data.