Update of the "data cleaning" part of the reconstruction.
We introduce a new DA  (MUONTRKOCCda, the version here is really a 1.0 one...) 
to compute the occupancy of the tracker (at the MANU level). This DA is supposed to run for all events of all PHYSICS runs.

The DA outputs an ASCII file onto the FXS. A new subprocessor (AliMUONOccupancySubprocessor) then reads this file
 (using an updated version of AliMUONTrackerIO) and upload an AliMUONVStore of the occupancy map
 to the OCDB (MUON/Calib/OccupancyMap).

The occupancy map might then be used in AliMUONPadStatusMaker (depending on new parameters in
 AliMUONRecoParam)  to reject some manus during the "calibration" stage of the reconstruction. 

Note that this new occupancy map supersedes the usage of the KillMap, which is now discontinued (and removed from the code and
 the svn OCDB).