Finally got the ResetDetector correct.
This required to insure that histograms are correctly named, in order to be able to distinguish Tracker from Trigger histograms.
Otherwise trigger part was resetting tracker part...

Also updated the shifter tracker occupancy plot to show DDL occupancy instead of buspatch occupancy. It is higher level and thus easier to explain and interpret.