Since there are some cuts which require to take informations from the event itself (event if they then cut on tracks or pairs), then I added a pointer to current AliRsnEvent which is assumed to be set up on it during the "LoopPair" method which makes the computations.
This required to propagate the corresponding setter in all classes which manage these cuts.

Moreover, I commented the "OpenFile" statement in AliRsnAnalysisSE, which seems to cause trouble
when running on PROOF clusters.

For ME analysis objects, I fixed a compilation error which appeared to me.

In AliRsnPair, I replaced static with dynamic casts to be more sure of how the histos are filled.