New analysis profiling implementation. Triggering collection of system info works only in "local" modeis done as usual via:    mgr->SetNSysInfo(nevents); // Dump info in sysinfo.txt each nevents A final displayed canvas contains: alocated VM[kB] per event for each top task (plot includes all subtasks), alocated VM for all tasks in CreateOutputObjects, alocated VM for all tasks in Terminate, VM profile for full analysis.
The memory profile is stored in a tree in a file sysinfo.root. One can investigate his task via:
   mgr->ProfileTask(const char *taskname); or:
   mgr->ProfileTask(Int_t itask)
where: taskname is the task name and itask is the task index in the list of top level tasks.
This will produce a canvas with memory alocation in CreateOutputObjects(), Exec() and Terminate() methods + CPU time spent per event