Major update of the way we make QA of the occupancy, and of what we present to the DQM shifter.
* DataMaker

Now at the level of raw data, the occupancy (OccRaws) is really that of raw data, i.e. if
one channel is on the data stream, it is counted (previously we first
redid the pedestal subtraction).

We also added the occupancy at the level of the calibrated digits (OccDigits)
(this is the same thing we had before at the raw level) and at the
recpoints level (OccRecPoints) (i.e. for only those digits that belong to a
reconstructed cluster). 

In principle, OccRaws > OccDigits > OccRecPoints.

Fixed a bug in the bin size of the buspatch occupancy. 

Adding accounting of empty events in the (previously known as readout
error plot) readout status plot, to make it easier to spot raw data corruption.

* DataChecker

Follow-up the change of signature of AliQACheckerBase::Check method,
so now checking (for Raws) is working again.

* DQM

We also added, for Raws, a plot with the DDL event size.
It is obviously related to the occupancy, but the
relationship between to two is not exactly straightforward.
That plot is replacing the DDL occupancy one for the shifter. 
So now the shifter sees Mean event size per DDL and a summary of the
readout errors (including the empty events).