new V0 Monitoring task (Markus Heide)
The monitoring task creates plots of the cut variables after different
cut
steps (0: just quality cuts, 1: invariant mass cut, 2: all cuts) and
maps the
quantities of tracks taken out by quality and V0 PID cuts in
corresponding
histograms.

In the AliTRDv0Info, the Armenteros-Polanski method was added and a few
graphical cuts adapted from the HFE V0 selection by Matus Kalisky.
The Kalman filter is in as a method, but not working properly yet (so
the cuts
on Chi2/NDF are by default not set).