Extending the functionality of the physics selection. Instead of providing a yes/no decision, it supports now several concurrent offline trigger selections. At present these are (defined in AliVEvent.h --> EOfflineTriggerTypes): 1) AliVEvent::kMB: Minimum bias trigger (corresponding to the old behavior), all detectors 2) AliVEvent::kMBNoTRD: Same as kMB, but TRD not read out (this is a separate online trigger class which became active now) 3) AliVEvent::kMUON: Muon trigger 4) AliVEvent::kHighMult: High-multiplicity trigger For all those the (usual) offline SPD or V0 selection is performed. The available offline triggers will be extended when needed.
To use this functionality you call for your task SetCollisionCandidates with a mask, composed 
from the above flags. E.g. to select MB and MUON events:
myTask->SetCollisionCandidates(AliVEvent::kMB | AliVEvent::kMUON)
Doing that your task is only called for the requested events. 

If in your UserExec you want to look at the decision of the physics selection, you can do that 
in the same way as before with
UInt_t mask = fInputHandler->IsEventSelected()
and in mask the bits corresponding to the above offline triggers are set.

The output bit mask of the physics selection is stored in the AOD event header, and therefore 
the AOD analysis can work in the same way as ESD analysis. I.e. you can use 
myTask->SetCollisionCandidates and fInputHandler->IsEventSelected() in the exact same way.

Andrei & Jan Fiete