AOD filling follows the new offline trigger schema. The method AliAODHandler::BeginEvent() resets filling of the AOD for the current event, while the ESD/AOD filters may set it back if the event passes their physics selection mask. All filters must add to their UserExec mehod lines like:   // Filters must explicitely enable AOD filling in their UserExec (AG)   if (!AliAnalysisManager::GetAnalysisManager()->GetOutputEventHandler()) AliFatal("Cannot run ESD filter without an output event handler");   AliAnalysisManager::GetAnalysisManager()->GetOutputEventHandler()->SetFillAOD(kTRUE);
Note: 
Changed the logic of AliAnalysisTaskSE::fOfflineTrigger for backward compatibility: if the task does not set it (stays 0) it means all events are selected. If the mask is non 0 it is AND-ed with the current event mask (as implemented with Jan-Fiete). 
Again, if no event selection cuts object (like AliPhysicsSelection) is used at the level of the input handler, all events are now selected.