new version of the V0 stuff for the RefMaker (Alex W.)
A new member of AliTRDtrackInfo::AliESDinfo added which has the v0pid info.
One can access it via AliTRDtrackInfo::AliESDinfo::GetV0pid(). In
addition the structure of the infoGen had to be changed. Now the first loop
is over the v0s, the second over the ESDtrackss. In case an ESDtrack has
a valid v0pid it is set, otherwise the values stay at 0 for all species.

RefMaker only depends on ESDinfo and not on Friends. Also some
modifications for the RefMakerNN. 

bug fixes:
1. TObjArrays like fTracksBarrel etc. are deleted now at the beginning
of the UserExec. Otherwise one gets v0pid values for tracks which do not
origin from a valid v0.

2. in the AliTRDv0info. Now there is a protection in case not both
tracks of a V0 are found (due to what reason ever).