Bastian Bathen:
* changed Configuration with new jet branch for gen

* New function UsePhysicsSelection(Bool_t):
Allows to ignore the physics selection (argument kFALSE), by default the physics selection is used. Since our task expects the physics selection running we were not able to switch it off so far.

* new histograms:
 - mean cross section (fh1Xsec) and sum of number of trials (fh1Trials) (both for weight of pt hard bins)
 - pt hard

* new function Notify() (copy from AliAnalysisTaskJetSpectrum):
 - is automatically called by analysis manager
 - fills fh1Xsec and fh1Trials

* change of (default) binning of some histograms, especially pt range (please check if new range is fine, I also modified ranges of IJ and DJ histograms)

* made the track and jet type constants ( enum{kTrackUndef=0, ...} ) public
(Since the constants were private in previous implementions they were not known in the AddTask. I do not understand why this problem did not occur earlier.)