Changes from Tomas: AliAnaElectron: Use positve track labels. Removed condition that the momentum vector should point to the emcal. Should be if only the track itself points to the emcal. B-field problem solved.
parmaker: in PWG4/macros/electrons is not working properly. Changed it
and added new script to make it work properly. New file is
makeparfilesproperly

makeparfilesproperly:  new bash script to make parfiles in PWG4/macros/electrons

AliAnaBtag: cleaned it up a bit and added btagging to tracks not only associated to the
geometry of the EMCAL.