Renamed script to add Central AOD task from AddTaskCentral.C to AddTaskCentralMult.C  - to be consistent.
Fixes for the event normalisation calculations.  We
can now properly use beta in the formula 

  N = 1/E_T (N_A + N_A/N_V (N_T - N_V - beta))

Note, that this requires that the physics selection 
does not check the that the hw trigger class is a 
collision trigger. 

Fixes for the TrainSetup.C script.  

MakeAOD.C and MakedNdeta.C will now use TrainSetup.C
if libRAliEn.so is available and working. 

Run.sh likewise will use the TrainSetup.C, unless 
libRAliEn.so is not available (checked) or an empty
name is passed via the option -N. 

DrawdNdeta.C fixed upd a little. 

Note, to check for a true collision, one now needs to 
do 
  
   AliAODForwardMult::IsTriggerBits(kB|kNSD)

for example - otherwise you also get A, C, AC, and E 
triggers that are also NSD off-line triggers.   This 
is done automatically in 

   AliAODForwardMult::CheckEvent