AliTRDdigitizer.cxx -> For the new name of the Nodata status in AliTRDCalChamberStatus
AliTRDcalibDB.cxx .h -> For the function related to the new status in AliTRDCalChamberStatus + the exb object with a protection against 100. I plan namely to put a default object with 100, which will indicate that the look up table as to be used. In AliTRDcalibDB, it takes the loop up take when the mean value is above 70...

AliTRDCalibChamberStatus.cxx .h -> calibration with tracks for chamber status of half chamber also

Cal/AliTRDCalChamberStatus.h .cxx -> change of definition of status, keeping the back compatibility.

AliTRDCalibTask.cxx .h -> Run the chamber status calibration + propagate the version and subersion number of the exb object used in the reconstruction

AliTRDCalibraFillHisto.cxx .h -> due to the new name of the status

AliTRDCaibraFit.cxx  -> some printf less