Major overhaul of the QA code.
The individual DrawXXX.C macros are almost all deprecated now.  Instead 
we use the QATrender and QAPlotter classes. 

The QATrender class produces a tree with entries per event of various QA
numbers.  This can be processed by QAPlotter to produce trend plots. 

The script RunQA.C will do both of this in one go. 
The shell script RunQA.sh can take care of this, and download the needed
files as well given the job id of the QA train. 

This should make it very easy to do the QA analysis.  Of course, one still
needs to apply a good part of subject judgement before making a statement
in the run condition table.