Updates for Grid analysis.
In the base class, the centrality bin objects where not re-created 
in Terminate.  This was added.  Sum objects are constructed in 
Terminate from the histograms in the first output list. 

If one has a data-member which isn't default constructable - 
say TH1*, then that pointer cannot be zero if the object is 
written to disk.  A simple example 

   struct Foo : public TObject
   {
     Foo(bool make=false) : fHist(0) 
     {
       if (!make) return;
       fHist  = new TH1D("h", "h", 30, -3, 3);
       fHist->FillRandom("gaus");
     }
     void Draw(Option_t* option="") 
     {
        if (fHist) fHist->DrawCopy();
     }
     TH1* fHist;
     static void Run(bool make=false)
     {
       TFile* out = TFile::Open("foo.root", "RECREATE");
       Foo*   oFoo = new Foo(make);
       oFoo->Write("foo");
       out->Close();
       
       TFile* in   = TFile::Open("foo.root", "READ");
       Foo*   iFoo = static_cast<Foo*>(in->Get("foo"));
       iFoo->Draw();
       in->Close();
     }
     ClassDef(Foo,1)
   };

If one does Foo::Run(false) - i.e., fHist == 0, then the write to disk failes. 
If one does Foo::Run(true) - i.e., fHist != 0, and of concrete type, all OK

So to have a null pointer, we need to specify explicitly the class of the 
data member fHist could be a TH1F* 

Therefore, the fShapeCorr must be of a TH2F object.