fix for bug https://savannah.cern.ch/bugs/?98881
Decouple the definition of the number of dEdx slices from their physics
meaning. In the TRD tracklet the definition represents allocated memory.
The memory is filled in AliTRDtrackV1::UpdateESDtrack according to
physics definitions from AliTRDPIDResponse and conventions.

The fix has to be suplemented by changes in AliESDtrack.cxx and
AliAODPid.h as requested in https://savannah.cern.ch/bugs/?98881#comment5