Chiara O. implemented a way to bypass HIJING internal calculation of the sigmaNN so that one can set the value from outside, and to switch off the elastic collisions. By default everything should be as before.
Constantin L.  then improved the calculation of the geometrical Npart,
which now is very close to the MC Glauber one. It still differs
because HIJING normalizes internally to collisions with the eikonal
approach for which the geometrical Npart can be zero (if we would
accept events for which the geometrical Npart > 0 then the result would be equal to the one of the MC Glauber).