Major refactoring of the code.
Introduced two base classes:

 * AliBaseESDTask:  Base class for tasks to read-in ESD data 
 * AliBaseAODTask:  Base class for tasks to read-in AOD data 

The various tasks has been moved to these base classes.  This is 
to avoid too much code duplication and to make the output more
consistent, as well as to insure that some diagnostics is 
produced. 

Note, these base classes defines a new interface for the sub-classs
to implement rather than the AliAnalysisTaskSE interface. 

The code has been tested throught the full chain and should be ready
to deploy. 

Also some other minor fixes for other code.