Temporary patch for Nano AOD tests in AliAODEvent
We are storing an AliNanoAODHeader in the Nano AODs, but all other AOD
classes still assume an AliAODHeader.  Calling methods from
AliAODHeader on an AliNanoAodHeader produces an undefined behavior. In
particular, AliAODEvent::ClearStd was calling fHeader->Clear(); Added
a protection to cast the header to AliVHeader in case it is not a
standard header.
If the Nano AOD solution is accepted, fHeader will be of AliVHEader.