Task and library for jet test in ESE spectra analysis (M Tangaro)
This new tasks builts the jet spectra using the standard PWG-JE
machinery in events with high and low q. This is used to evaluate
possible biases from hard processes in the computation of the Q
vector.  It is built as a separate lib because it depends on the Jet
libraries (building it in libPWGspectra would have introduced a
dependency on the latter).