.so cleanup: restored .so in SetAdditionalLibs()
gSystem->Load() does not need .so, .dylib extension: ROOT picks one
automatically; extensions there should *not* be specified in order to
make the code multiplatform.

Analysis Plugin's SetAdditionalLibs() *requires* libraries to have
either the ".so" or the ".dylib" extension because it needs to
distinguish libraries from PARfiles, headers, etc.: one of ".so" or
".dylib" *must* be specified, and at the time of loading it will be
correctly stripped - meaning that, for instance, "libASD.so" will work
on Macs too and "libASD.dylib" will also work on Linux boxes.

So:

 - gSystem->Load( "libASD" )  // no extension!
 - plugin->SetAdditionalLibs( "libASD1.so libASD2.dylib ..." )
     // must use either .so or .dylib