Common driver for tests
Introduces the test/runTests helper script and the test/test.cfg
configuration, which mimicks the behavior of the various runtest.sh
files.

By default running:

  test/runTests <test-name>

should be exactly the same as:

  cd <test-name> && ./runtest.sh

Besides some level of cleanup coming from avoiding repetition, this also
introduces so called "variants" which can be specified so that different
kind of tools can be used to run the same test, allowing to configure
the output name depending on the variant, avoiding having to repeat the
command. An example is the variant `igprof_memory` which will prefix the
specified test with the commands required to run igprof.