Add support for multiple outputs to alihadd
By using the -s option of alihadd, the user can specify a maximum "cost"
for the output file which if surpassed will force the creation of an
additional output file.

Right now the cost is calculated as the sum of the size of the TKeys to be
merged. Clearly this is not a good cost for histogram merges, but it's quick
and good enough for now.