Usage of reference correction map implemented
Added loading of the reference correction map from TPC/Calib/CorrectionMapsRef on top of the
run-dependent TPC/Calib/CorrectionMap.
It should be prepared as an TObjArray of maps per field polarity, with corresponding polarity flag set.
If RecoParam->GetUseDistortionFractionAsErrorY() and GetUseDistortionFractionAsErrorZ() are non-zero,
the cluster will get addtinal error assigned as squared corresponding fraction for difference between the time-dependent
and reference maps