Adding option for decay channels D0->pi0 pi+ pi- and D0->K- pi+ pi0
Decay mode added as extra option and is switched off
for existing targets, so there is no change of behaviour
for the existing settings. New targets including "PionicD0"
in the name are added in order to switch on the pionic D0
decay mode.

Multiple options for the pionic decay modes are added in order
to distinguish forcing the decay channels:
- 0: Pionic channels swiched off
- 1: K- pi+ and pionic channels enabled
- 2: Only pionic channels
- 3: Only D0 - > K- pi+ pi0
- 4:  Only D0 -> pi0 pi+ pi-
In addition to the existing pionic modes additional
enums values for the pure pionic modes are added to
the AliDecayer.h.