Port MC option for Xic0 semileptonic mode  toV5 09 54 patches (#1478)
* switch on D and B decays

* Fix for hypernuclei masses

The masses of the hyper hydrogen4 and the hyper helium4 need to be corrected since in the old implementation only the lambda separation energy of 2MeV/c2 was substracted from the sum of the constituents. This is incorrect since also the binding energy of the remaining nucleus (triton and he3 respectively) needs to be taken into account. Furthermore there is the existing of an excited state for both particles (see A. Gal, E.V. Hungerford and D.J. Millener, „Strangeness in nuclear physics“, Rev. Mod. Phys. 88 (2016) 3, 035004, arXiv:1605.00557 [nucl-th]) which is implemented now with the correct mass after the (very short lived) EM transition to the ground state (decay modes then implemented as for the ground state).

* Fix for hypernuclei masses

The masses of the hyper hydrogen4 and the hyper helium4 need to be corrected since in the old implementation only the lambda separation energy of 2MeV/c2 was substracted from the sum of the constituents. This is incorrect since also the binding energy of the remaining nucleus (triton and he3 respectively) needs to be taken into account. Furthermore there is the existing of an excited state for both particles (see A. Gal, E.V. Hungerford and D.J. Millener, „Strangeness in nuclear physics“, Rev. Mod. Phys. 88 (2016) 3, 035004, arXiv:1605.00557 [nucl-th]) which is implemented now with the correct mass after the (very short lived) EM transition to the ground state.

* Including (anti)4Xi(-)He (#1442)

* Including (anti)4Xi(-)He

Following Gal et al. Phys.Lett.B 820 (2021) 136555 for the binding energy and using the free Xi- lifetime from the pdg. Applying the pdg numberig scheme for nuclei/hypernuclei, but changing the fixed "10" in the beginning to "12" for Xi hypernuclei (leaving "11" for potential Sigma hypernuclei).

* Including second decay channel for (anti)4Xi(-)He

* Update Hypernuclei Masses

Update of the hypernuclei masses taken from the hypernuclei database (https://hypernuclei.kph.uni-mainz.de/). High precission of the masses is necessary for the hypertriton.

* Update hypernuclei masses and including 4Xi(-)He

Update hypernuclei masses according to the hypernuclei database (https://hypernuclei.kph.uni-mainz.de/) and including 4Xi(-)He following Gal et al. Phys.Lett.B 820 (2021) 136555 for the binding energy and using the free Xi- lifetime from the pdg. Applying the pdg numberig scheme for nuclei/hypernuclei, but changing the fixed "10" in the beginning to "12" for Xi hypernuclei (leaving "11" for potential Sigma hypernuclei).

* Update for 5LHe

Fix Typo in 5LHe Mass

* Update for 5LHe

Fix Typo

* Update on mass A=4 hypernuclei MC (#1454)

* Update on mass A=4 hypernuclei
* Update H-Dibaryon (Resonance) and OmegaP bound state

* Update Xic0 lifetime (#1471)

* MC option for Xic0 semileptonic mode (#1452)

* add 4 body decay modes for kXic0Semileptonic
* Add the 4body decay mode for the Xic0 semileptonic simulation

---------

Co-authored-by: Andreas Morsch <andreas.morsch@cern.ch>
Co-authored-by: jditzelnew <58816213+jditzelnew@users.noreply.github.com>
Co-authored-by: arossi81 <andrea.rossi@cern.ch>
Co-authored-by: Sanghoon Lim <shlim1986@gmail.com>